requirements.txt
h5py==3.10.0
numpy==2.1.0
opencv_python==4.9.0.80
pandas==2.2.2
pytorch_msssim==1.0.0
PyYAML==6.0.2
scipy==1.14.1
setuptools==59.6.0
skimage==0.0
tifffile==2024.2.12
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
